refactor: extract useCharacterState hook from AppContext
Move 9 character/avatar useState calls + 6 callbacks (loadCharacter,
handleSaveCharacter, handleCharacterFieldInput, handleCharacterArrayInput,
handleCharacterStyleInput, handleCharacterMessageExamplesInput) plus
the selectedVrmIndex localStorage+server sync wrapper.

Cross-domain coupling: hook accepts { agentStatus, setAgentStatus } params
for the character-save → agent-name-update bridge.

AppContext.tsx: 8,150 → 8,048 lines (-102 from this extraction)
useCharacterState.ts: 222 lines (new)
